Split out flight plan test.
